refactor: vdom
