+++ src/vdom/mailbox.rs
@@ -1,0 +1,25 @@
+use std::rc::Rc;
+
+pub struct Mailbox<Message: 'static> {
+    func: Rc<dyn Fn(Message)>,
+}
+
+impl<Ms> Mailbox<Ms> {
+    pub fn new(func: impl Fn(Ms) + 'static) -> Self {
+        Mailbox {
+            func: Rc::new(func),
+        }
+    }
+
+    pub fn send(&self, message: Ms) {
+        (self.func)(message)
+    }
+}
+
+impl<Ms> Clone for Mailbox<Ms> {
+    fn clone(&self) -> Self {
+        Mailbox {
+            func: self.func.clone(),
+        }
+    }
+}
